requirements.txt
Django>=5.2.4,<6.0
Pillow==11.3.0
django-admin-sortable2==2.2.8